drivers: input: nt36xxx: fix pointer casting

+++ drivers/input/touchscreen/nt36xxx_spi/nt36xxx_ext_proc.c
@@ -280,7 +280,7 @@ return:
 *******************************************************/
 static void *c_start(struct seq_file *m, loff_t *pos)
 {
-	return *pos < 1 ? (void *)1 : NULL;
+	return *pos < 1 ? (void *)(unsigned long)1 : NULL;
 }
 
 /*******************************************************
